Declare package as automatic
Prior to this commit, the lifetime of all functions in this package
defaulted to *static*, meaning their arguments and local variables kept
their value between function invocations, which was unintended.

This commit changes the lifetime to all functions in this package to
*automatic*, meaning function arguments and local variables are
allocated upon entry into the function and deallocated upon exit.